Move `injectOEmbedMarkup` function to the consumer
