+++ packages/slate-editor/src/components/HtmlInjection.tsx
@@ -3,8 +3,6 @@ import React, { useEffect, useRef } from 'react';
 
 import { useLatest } from '#lib';
 
-import { injectOembedMarkup } from '#modules/editor-v4-embed/lib';
-
 interface Props {
     html: string;
     className?: string;
@@ -28,3 +26,38 @@ export const HtmlInjection: FunctionComponent<Props> = (props) => {
 
     return <div className={className} ref={ref} />;
 };
+
+export function injectOembedMarkup({
+    html,
+    onError,
+    target,
+}: {
+    html: string | undefined;
+    onError: () => void;
+    target: HTMLElement;
+}): void {
+    const container = document.createElement('div');
+    container.innerHTML = html || '';
+    const embedScripts = Array.from(container.getElementsByTagName('script'));
+
+    embedScripts.forEach((embedScript) => {
+        const script = document.createElement('script');
+        copyScriptAttributes(embedScript, script);
+        script.addEventListener('error', onError);
+
+        document.body.appendChild(script);
+        // Remove the original script so it's not loaded twice.
+        embedScript.remove();
+    });
+
+    // eslint-disable-next-line no-param-reassign
+    target.innerHTML = container.innerHTML;
+}
+
+function copyScriptAttributes(source: HTMLScriptElement, target: HTMLScriptElement) {
+    Array.from(source.attributes).forEach(({ name, value }) => {
+        target.setAttribute(name, value);
+    });
+    // eslint-disable-next-line no-param-reassign
+    target.innerText = source.innerText;
+}
