+++ packages/slate-editor/src/modules/editor-v4-embed/lib/index.ts
@@ -1,5 +1,4 @@
 export { createEmbed } from './createEmbed';
-export { injectOembedMarkup } from './injectOembedMarkup';
 export { normalizeRedundantEmbedAttributes } from './normalizeRedundantEmbedAttributes';
 export { parseSerializedElement } from './parseSerializedElement';
 export { useFloatingEmbedInput } from './useFloatingEmbedInput';
